+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Updated Bugsnag library and exception handling to v2 API
+- [launcher] Improve robustness
 
 ## [1.0.2] - 2018-04-09